Fix black screen, add test

+++ gazebo/gui/model/ModelCreator.cc
@@ -65,10 +65,11 @@ ModelCreator::ModelCreator()
 
   this->jointMaker = new JointMaker();
 
-  connect(g_deleteAct, SIGNAL(DeleteSignal(const std::string &)), this,
-          SLOT(OnDelete(const std::string &)));
-
-  this->Reset();
+  if (g_deleteAct)
+  {
+    connect(g_deleteAct, SIGNAL(DeleteSignal(const std::string &)), this,
+            SLOT(OnDelete(const std::string &)));
+  }
 }
 
 /////////////////////////////////////////////////
@@ -94,6 +95,11 @@ std::string ModelCreator::CreateModel()
 /////////////////////////////////////////////////
 void ModelCreator::AddJoint(JointMaker::JointType _type)
 {
+  if (!this->modelVisual)
+  {
+    this->Reset();
+  }
+
   this->Stop();
   if (this->jointMaker)
     this->jointMaker->CreateJoint(_type);
